Adding Bandcamp URL/API Parser, updating README with usage.
The Bandcamp parser acts on user-facing URLs. It is not the intention
to force you to put together an API URL. A whole slew of new modules
are required to execute, as listed in the README. Your API key must be
added into your environment under the BCAPIKEY variable. This should be
mostly cross platform, able to be done globally (/etc/profile), locally
(~/.bash_profile, ~/.zshrc, etc.), and even under Windows (I don't
remember where the env dialog is).